some refactoring to increase code climate
